Simplify repo to keep only the seq_embedded model and consolidate shared utilities.
- Consolidated `models.py`, `data.py`, `trainer.py`, `service.py`, and `deploy.py` at the top level.
- Flattened configuration and model classes, removing unnecessary base classes and inheritance.
- Removed `mf/`, `seq/`, and `seq_embedded/` subdirectories.
- Added unit tests in `tests/test_recommender.py` for data loading, model forward pass, and training/validation steps.
- Updated `CLAUDE.md`, `README.md`, and `pyproject.toml` to reflect the new structure and entry points.
- Verified changes with `pytest` and basic training/deployment smoke tests.
