--- xfmr_rec/mf/__init__.py
@@ -1,1 +1,0 @@
-MODEL_NAME = "xfmr_mf_rec"
