--- xfmr_rec/seq/__init__.py
@@ -1,1 +1,0 @@
-MODEL_NAME = "xfmr_seq_rec"
