--- xfmr_rec/seq_embedded/__init__.py
@@ -1,1 +1,0 @@
-MODEL_NAME = "xfmr_seq_embedded_rec"
